+++ packages/browser-wallet/src/popup/popupX/constants/routes.ts
@@ -222,6 +222,12 @@ export const relativeRoutes = {
                 backTitle: '',
             },
         },
+        configuration: {
+            path: 'configuration',
+            config: {
+                backTitle: '',
+            },
+        },
         passcode: {
             path: 'passcode',
         },