+++ packages/browser-wallet/src/popup/popupX/page-layouts/MainLayout/Header/components/MenuTiles.tsx
@@ -10,6 +10,7 @@ import Info from '@assets/svgX/info2.svg';
 import Restore from '@assets/svgX/arrow-counter-clock.svg';
 import Eye from '@assets/svgX/eye.svg';
 import NFT from '@assets/svgX/cube-focus.svg';
+import Gear from '@assets/svgX/gear.svg';
 import IconButton from '@popup/shared/IconButton';
 import Text from '@popup/popupX/shared/Text';
 import { Link } from 'react-router-dom';
@@ -38,21 +39,27 @@ export default function MenuTiles({ menuOpen, setMenuOpen }: MenuTilesProps) {
                 onClick={() => setMenuOpen(false)}
             >
                 <Link to={absoluteRoutes.settings.identities.path}>
-                    <IconButton className="main-header__menu-tiles_tile wide">
+                    <IconButton className="main-header__menu-tiles_tile">
                         <Identification />
                         <Text.Capture>{t('identities')}</Text.Capture>
-                    </IconButton>
-                </Link>
-                <Link to={absoluteRoutes.settings.accounts.path}>
-                    <IconButton className="main-header__menu-tiles_tile wide">
-                        <Browsers />
-                        <Text.Capture>{t('accounts')}</Text.Capture>
                     </IconButton>
                 </Link>
                 <Link to={absoluteRoutes.settings.seedPhrase.path}>
                     <IconButton className="main-header__menu-tiles_tile">
                         <TextColumns />
                         <Text.Capture>{t('seedPhrase')}</Text.Capture>
+                    </IconButton>
+                </Link>
+                <Link to={absoluteRoutes.settings.accounts.path}>
+                    <IconButton className="main-header__menu-tiles_tile">
+                        <Browsers />
+                        <Text.Capture>{t('accounts')}</Text.Capture>
+                    </IconButton>
+                </Link>
+                <Link to={absoluteRoutes.settings.configuration.path}>
+                    <IconButton className="main-header__menu-tiles_tile">
+                        <Gear />
+                        <Text.Capture>{t('settings')}</Text.Capture>
                     </IconButton>
                 </Link>
                 <Link to={absoluteRoutes.settings.passcode.path}>