+++ packages/browser-wallet/src/popup/shell/i18n/locales/en.ts
@@ -42,6 +42,7 @@ import restore from '@popup/popupX/pages/Restore/i18n/en';
 import connectedSites from '@popup/popupX/pages/ConnectedSites/i18n/en';
 import privateKey from '@popup/popupX/pages/PrivateKey/i18n/en';
 import seedPhrase from '@popup/popupX/pages/SeedPhrase/i18n/en';
+import configuration from '@popup/popupX/pages/Settings/i18n/en';
 import passcode from '@popup/popupX/pages/ChangePasscode/i18n/en';
 import network from '@popup/popupX/pages/NetworkSettings/i18n/en';
 import connect from '@popup/popupX/pages/ConnectNetwork/i18n/en';
@@ -109,6 +110,7 @@ const t = {
         connectedSites,
         privateKey,
         seedPhrase,
+        configuration,
         passcode,
         network,
         connect,