+++ packages/browser-wallet/src/shared/analytics/index.ts
@@ -1,0 +1,1 @@
+export { appTracker as default } from './app-tracker';